dump.f2fs: Dump symlinks as symlinks
Previously, dumped symlinks would always create regular files instead.
This allows symlinks to be dumped as symlinks with the -L option.

The i_name field's name may not be the same as the actual name from the
dirent, so we use the dirent name when available.

Currently hardlinks aren't detected, so print a warning if we notice a
nondirectory with a link count over 1.

Test: atest CopyEfsTest
Flag: TEST_ONLY
Change-Id: I9c25bcdd44274108b5720f317b187515f64c3622
Signed-off-by: Daniel Rosenberg <drosen@google.com>